refactor(backend): separating concerns for auth files and spotifyAPI class
